混雑度口コミ表示のカードをコンポーネント化#11
